refactored screen stacks into routes directory in files Authenticated.js and SignUp.js and setup base for authentication flow   Need to implement authenticated token using secure store
