+++ App.js
@@ -1,132 +1,20 @@
-import LandingScreen from "./app/screens/LandingScreen";
-import LoginScreen from "./app/screens/LoginScreen";
-import ForgotLoginScreen from "./app/screens/ForgotLoginScreen"
-import SignUpScreen from "./app/screens/SignUpScreen";
-import SignUpScreenMore from "./app/screens/SignUpScreenMore";
-import HomeScreen from "./app/screens/HomeScreen";
-import PeersScreen from "./app/screens/Peers";
-import ParentsScreen from "./app/screens/Parents";
-import ProfessionalsScreen from "./app/screens/Professionals";
-
-import Header from "./app/components/Header";
-import {Colors} from "./app/components/Colors";
-
-/*import SplashScreen from "./app/screens/SplashScreen";*/
-
-
 import AppLoading from "expo-app-loading";
 import { Asset } from "expo-asset";
 import Constants from "expo-constants";
 import * as SplashScreen from "expo-splash-screen";
 import { useCallback, useEffect, useMemo, useState } from "react";
-
-
 import {NavigationContainer} from "@react-navigation/native";
-import {createNativeStackNavigator} from "@react-navigation/native-stack";
 
-import { createMaterialTopTabNavigator } from '@react-navigation/material-top-tabs';
 import {StyleSheet, Text, TextComponent, Animated, Button, Platform, View,} from "react-native";
 
-const Stack = createNativeStackNavigator();
-const Tab = createMaterialTopTabNavigator();
+//avigators
+import SignInNavigator from './app/routes/SignUp'
+import AuthenticatedNavigator from "./app/routes/Authenticated";
+
 
 SplashScreen.preventAutoHideAsync().catch(() => {
     /* reloading the app might trigger some race conditions, ignore them */
 });
-
-function LoginSignIn() {
-    return (
-        <Tab.Navigator style={styles.tabItems}
-                       screenOptions={{
-                           //tabBarStyle: { backgroundColor: 'red'},
-                           //tabBarActiveTintColor: 'red',
-                           /*               indicatorStyle: {
-                                              backgroundColor: "red",
-                                          }*/
-
-                       }}
-        >
-            <Tab.Screen name="Login" component={LoginScreen} />
-            <Tab.Screen name="Sign Up" component={SignUpScreen} />
-        </Tab.Navigator>
-    );
-}
-
-function AuthenticatedStack() {
-    return (
-        <Stack.Navigator >
-            <Stack.Group>
-                <Stack.Screen
-                    name="Home"
-                    component={HomeScreen}
-                />
-                <Stack.Group>
-                    <Stack.Screen
-                        name="Peers"
-                        component={PeersScreen}
-                        options={ ( {navigation}) => {
-                            return {
-                                headerTitle: () =>
-                                    <Header
-                                        title="Peers"
-                                        left="Professionals"
-                                        right="Parents"
-                                        navigation={navigation}
-                                    />,
-                                headerBackVisible: false,
-                                headerStyle: {
-                                    backgroundColor: Colors.secondaryGreen,
-                                }
-                            }
-                        }}
-                    />
-                    <Stack.Screen
-                        name="Parents"
-                        component={ParentsScreen}
-                        options={ ( {navigation}) => {
-                            return {
-                                headerTitle: () =>
-                                    <Header
-                                        title="Parents"
-                                        left="Peers"
-                                        right="Professionals"
-                                        navigation={navigation}
-                                    />,
-                                headerBackVisible: false,
-                                headerStyle: {
-                                    backgroundColor: Colors.primaryBlue,
-                                }
-                            }
-                        }}
-                    />
-                    <Stack.Screen
-                        name="Professionals"
-                        component={ProfessionalsScreen}
-                        options={ ( {navigation}) => {
-                            return {
-                                headerTitle: () =>
-                                    <Header
-                                        title="Professionals"
-                                        left="Parents"
-                                        right="Peers"
-                                        navigation={navigation}
-                                    />,
-                                headerBackVisible: false,
-                                headerStyle: {
-                                    backgroundColor: Colors.primaryRed,
-                                }
-                            }
-                        }}
-                    />
-                </Stack.Group>
-            </Stack.Group>
-
-        </Stack.Navigator>
-    )
-}
-
-//TODO authenticaiton workflow
-/*Authenticated ? AuthenticatedStack : LoginSignIn*/
 
 function AnimatedAppLoader({ children, image }) {
     const [isSplashReady, setSplashReady] = useState(false);
@@ -218,63 +106,23 @@ function AnimatedSplashScreen({ children, image }) {
         </View>
     );
 }
+//TODO authentication workflow  replace authenticated with secure store token
 
-export default function App() {
-    return (
+let Authenticated = false;
 
-        <AnimatedAppLoader image={require('./app/assets/photoboom_logo.png')}>
-
-            <NavigationContainer>
-                <Stack.Navigator>
-                    <Stack.Screen
-                        name="SignIn"
-                        //component={LoginSignIn}
-                        component={AuthenticatedStack}
-                        //component={SplashScreen}
-                        //component={SignUpScreenMore}
-                        //options={{title:'PhotoBoom'}}
-                        //options={{title:''}}
-                        options={{ headerShown: false, marginTop: 20}}
-                    />
-                    <Stack.Screen
-                        name="SignUpMore"
-                        component={SignUpScreenMore}
-                        options={({ route }) => ({title: route.params.u_email})}
-                    />
-                    <Stack.Screen
-                        name={"Landing"}
-                        component={LandingScreen}
-
-                        options={({ route }) => ({
-                            headerTitle:"Photoboom  ",
-                            headerStyle: {
-                                backgroundColor: '#008000',
-                            },
-                            headerTintColor: '#fff',
-                            headerTitleStyle: {
-                                fontWeight: 'bold',
-                                fontStyle: 'italic',
-                                textAlign: 'center',
-                            },
-                        })}
-                    />
-                    <Stack.Screen
-                        name="Forgot"
-                        component={ForgotLoginScreen}
-                    />
-                </Stack.Navigator>
-            </NavigationContainer>
-        </AnimatedAppLoader>
-    )
-}
-
-
-
-const styles = StyleSheet.create({
-    tabItems: {
-        //paddingTop: 10,
-        backgroundColor: "red",
-        marginTop: 60,
+    export default function App() {
+        return (
+            <AnimatedAppLoader image={require('./app/assets/photoboom_logo.png')}>
+                <NavigationContainer>
+                    {Authenticated == false ? (
+                        <SignInNavigator />
+                        ) : (
+                        <AuthenticatedNavigator />
+                    )}
+                </NavigationContainer>
+            </AnimatedAppLoader>
+        )
     }
 
-});
+
+
